chore(libknossos): address go linter warnings

+++ packages/server/cmd/importer/main.go
@@ -369,7 +369,7 @@ TRUNCATE files CASCADE;
 
 				fid, err := q.CreateExternalFile(ctx, queries.CreateExternalFileParams{
 					StorageKey: "ext#" + nanoid.New(),
-					Filesize:   int32(archive.FileSize),
+					Filesize:   archive.FileSize,
 					Public:     true,
 					External:   archive.URLs,
 					Owner:      1,
